+++ readFiles.h
@@ -11,12 +11,12 @@
 
 using namespace std;
 
-void readSites(const string &sitefname, vector<vector<double> > &sites );
+void readSites(const string &sitefname, vector<vector<int> > &sites );
 void readLocs(const string &locfname, vector<int> &locs );
 void readHapInfo(const string &fname, vector<vector<string> > &hapInfo );
 
 void print1Dvec(const vector<int>& vec);
-void print2Dvec(const vector<vector<double> >& vec);
+void print2Dvec(const vector<vector<int> >& vec);
 void print2DvecString(const vector<vector<string> >& vec);
